trading_bot/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
boto3>=1.28.0
pyyaml>=6.0
alpaca-trade-api>=3.0.0
requests>=2.31.0
pytest>=7.4.0
pytest-cov>=4.1.0
moto>=4.2.0
